Logic for generating screenshot filename
I'm confused about the logic here, if the `fileName` is `abc.ext` it's asking if `abc` equals `.png`. That's not right, right?

Here I've swapped in `Path.GetExtension`, but I'm new to C# and I can't run the test suite of this package so I don't know exactly what the right solution is. Any thoughts?